Challenge
Challenge
One of the best ways to learn is to write your own code as often as possible, so here are three ways you should try your new knowledge to make sure you fully understand what’s going on:

For an easy challenge try adding a score label that updates as the player’s score changes.
Make the game end after a certain number of launches. You will need to use the invalidate() method of Timer to stop it from repeating.
Use the waitForDuration and removeFromParent actions in a sequence to make sure explosion particle emitters are removed from the game scene when they are finished.